DX-20021: Updating documents when version is null

+++ core/src/main/java/site/ycsb/workloads/RepeatedUpdatesWorkload.java
@@ -77,7 +77,7 @@ public class RepeatedUpdatesWorkload extends Workload {
     final HashMap<String, ByteIterator> updatedValues = new HashMap<>();
     // JobStates are numeric values. Iterate through job states.
     Status status = Status.OK;
-    Long version = null;
+    Integer version = null;
 
     for (String jobId: tracker.jobIds) {
       int i = 1;
@@ -85,7 +85,7 @@ public class RepeatedUpdatesWorkload extends Workload {
         updatedValues.put("jobState", new StringByteIterator(getJobStateString(i)));
         Pair statusVersionPair = db.findAndUpdate("jobs", jobId, version,
             updatedValues);
-        version = (Long) statusVersionPair.getVersion();
+        version = (Integer) statusVersionPair.getVersion();
         status = statusVersionPair.getStatus();
         i += 1;
       }
